csv header

+++ workflow/scripts/rfAfterPhyML.py
@@ -14,10 +14,11 @@ def addToCsv(window_pos, ref_feature, normalized_RF,output_csv):
 def filter_RF(rf_outfile,output_csv):
     time.sleep(5)
     columns_name = ["window_pos", "ref_feature","normalized_RF"]
-    with open(output_csv, 'w') as f_object:
-        writer_object = writer(f_object)
-        writer_object.writerow(columns_name)
-        f_object.close()
+    if not os.path.exists(output_csv):
+        with open(output_csv, 'w') as f_object:
+            writer_object = writer(f_object)
+            writer_object.writerow(columns_name)
+            f_object.close()
     for the_file in rf_outfile:
         if os.stat(the_file).st_size == 0: 
             normalized_RF = 999
@@ -25,7 +26,7 @@ def filter_RF(rf_outfile,output_csv):
             with open(the_file, 'r') as file:
                 normalized_rf = file.read().rstrip()
             normalized_RF = round(float(normalized_rf)*100,2)
-        window_pos, ref_feature, = the_file[11:-7].split(".")
+        window_pos, ref_feature, = the_file.split("/")[-1].split(".")[:2]
         addToCsv(window_pos, ref_feature, normalized_RF,output_csv)
 
 if __name__ == '__main__':
